// src/lib/db-types.ts
export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Account {
  accessToken: string | null;
  accessTokenExpiresAt: Timestamp | null;
  accountId: string;
  createdAt: Generated<Timestamp>;
  id: string;
  idToken: string | null;
  password: string | null;
  providerId: string;
  refreshToken: string | null;
  refreshTokenExpiresAt: Timestamp | null;
  scope: string | null;
  updatedAt: Timestamp;
  userId: string;
}

export interface Batch {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  phase_progress_id: string;
  progress: Generated<number>;
  quantity: Generated<number>;
  user_id: string;
}

export interface Employee {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  name: string;
  shift_id: string | null;
  user_id: string;
}

export interface EmployeeSkill {
  employee_id: string;
  skill_id: string;
  user_id: string;
}

export interface Phase {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  max_batch: number | null;
  name: string;
  setup_time: Generated<number>;
  unit_time: Generated<number>;
  user_id: string;
}

export interface PhaseProgress {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  phase_id: string;
  product_order_item_id: string;
  sort_order: Generated<number>;
  user_id: string;
}

export interface PhaseSupply {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  phase_id: string;
  quantity: Generated<number>;
  supply_id: string;
  user_id: string;
}

export interface PhaseTool {
  count: Generated<number>;
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  phase_id: string;
  tool_id: string;
  user_id: string;
}

export interface PhaseWorker {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  phase_id: string;
  user_id: string;
}

export interface PhaseWorkerSkill {
  phase_worker_id: string;
  skill_id: string;
  user_id: string;
}

export interface Product {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  name: string;
  user_id: string;
}

export interface ProductOrder {
  created_at: Generated<Timestamp>;
  due_date: Timestamp | null;
  id: Generated<string>;
  name: string;
  order_date: Generated<Timestamp>;
  user_id: string;
}

export interface ProductOrderItem {
  count: Generated<number>;
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  product_id: string;
  product_order_id: string;
  user_id: string;
}

export interface ProductPhase {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  phase_id: string;
  product_id: string;
  sort_order: Generated<number>;
  user_id: string;
}

export interface Schedule {
  created_at: Generated<Timestamp>;
  date: Timestamp;
  id: Generated<string>;
  user_id: string;
}

export interface ScheduleShift {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  schedule_id: string;
  shift_id: string;
  user_id: string;
}

export interface Session {
  createdAt: Generated<Timestamp>;
  expiresAt: Timestamp;
  id: string;
  ipAddress: string | null;
  token: string;
  updatedAt: Timestamp;
  userAgent: string | null;
  userId: string;
}

export interface Shift {
  created_at: Generated<Timestamp>;
  end_time: string;
  id: Generated<string>;
  name: string;
  start_time: string;
  user_id: string;
}

export interface Skill {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  name: string;
  user_id: string;
}

export interface Supply {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  name: string;
  quantity: Generated<number>;
  user_id: string;
}

export interface SupplyOrder {
  arrival_date: Timestamp | null;
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  name: string;
  order_date: Generated<Timestamp>;
  user_id: string;
}

export interface SupplyOrderItem {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  quantity: Generated<number>;
  supply_id: string;
  supply_order_id: string;
  user_id: string;
}

export interface Tool {
  count: Generated<number>;
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  name: string;
  user_id: string;
}

export interface User {
  createdAt: Generated<Timestamp>;
  email: string;
  emailVerified: boolean;
  id: string;
  image: string | null;
  name: string;
  updatedAt: Generated<Timestamp>;
}

export interface Verification {
  createdAt: Generated<Timestamp>;
  expiresAt: Timestamp;
  id: string;
  identifier: string;
  updatedAt: Generated<Timestamp>;
  value: string;
}

export interface Work {
  batch_id: string;
  created_at: Generated<Timestamp>;
  employee_id: string;
  end_time: string;
  id: Generated<string>;
  schedule_shift_id: string;
  start_time: string;
  user_id: string;
}

export interface DB {
  account: Account;
  batch: Batch;
  employee: Employee;
  employee_skill: EmployeeSkill;
  phase: Phase;
  phase_progress: PhaseProgress;
  phase_supply: PhaseSupply;
  phase_tool: PhaseTool;
  phase_worker: PhaseWorker;
  phase_worker_skill: PhaseWorkerSkill;
  product: Product;
  product_order: ProductOrder;
  product_order_item: ProductOrderItem;
  product_phase: ProductPhase;
  schedule: Schedule;
  schedule_shift: ScheduleShift;
  session: Session;
  shift: Shift;
  skill: Skill;
  supply: Supply;
  supply_order: SupplyOrder;
  supply_order_item: SupplyOrderItem;
  tool: Tool;
  user: User;
  verification: Verification;
  work: Work;
}